Scattering rendering pass runs without exceptions
Buffer mapping sentinel object had to be refactored to unmap the buffer each
time after reading/writing access has been made from the host side as OptiX does not
support GPU-CPU synchronization primitives and buffers thus cannot be mapped to the pinned
memory persistently


Former-commit-id: 9b472f2a1dd7d5763d404acb0b2b442d2904acee

+++ ox_wrapper/ox_wrapper/ray_casters/parallel_ray_generator.cpp
@@ -17,7 +17,7 @@ OxParallelRayGenerator::OxParallelRayGenerator(OxContext const& context, uint32_
 {
     setGeneratorDimensions(num_rays, 1U, 1U);
     getRayGenerationShader().setVariableValue("num_rays", num_rays);
-    getRayGenerationShader().setVariableValue("num_spectra_pairs_supported", num_spectra_pairs_supported);
+    getRayGenerationShader().setVariableValue("num_spectra_pairs_supported", static_cast<uint32_t>(num_spectra_pairs_supported));
     setEmitterSize(emitter_size);
     setEmitterPosition(emitter_position);
     setEmitterRotation(emitter_rotation);
